requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
httpx>=0.27.0
firebase-admin>=6.5.0
python-dotenv>=1.0.1
pytz>=2024.1
itsdangerous>=2.2.0
slowapi>=0.1.9
